Move BumpType to src/lib.rs
