+++ src/lock/mod.rs
@@ -3,7 +3,7 @@ mod dump;
 mod parse;
 mod update;
 
-pub use bump::{git_bump, simple_bump, BumpType};
+pub use bump::{git_bump, simple_bump};
 pub use dump::dump;
 pub use parse::{parse, parse_git_line, parse_line, parse_simple_line};
 pub use update::Updater;
