+++ src/operations/lock.rs
@@ -4,10 +4,10 @@ use std::path::Path;
 use crate::dependencies::FrozenDependency;
 use crate::error::*;
 use crate::lock;
-use crate::lock::BumpType;
 use crate::lock::Updater;
 use crate::lock::{git_bump, simple_bump};
 use crate::ui::*;
+use crate::BumpType;
 use crate::Metadata;
 
 #[derive(Default, Debug)]
